rename index files for clearer import statements
