+++ src/components/ImportedMode/styles/uploader.css
