+++ src/components/QueryEditor/index.js
