+++ src/components/QuerySelector/index.js
