+++ src/components/SchemaDisplay/index.js
