+++ src/database/queryHistory.json
@@ -2,8 +2,8 @@
   {
     "queryText": "query writtenQuery(\n  $id: Int\n) {\n  Media(id: $id, type: ANIME) {\n    _id: id\n    title {\n      native\n      \n    }\n  }\n}\n",
     "url": "https://graphql.anilist.co",
-    "timeStamp": "2:41:07 PM",
-    "dateStamp": "8/17/2021"
+    "timeStamp": "1:52:56 PM",
+    "dateStamp": "8/18/2021"
   },
   {
     "queryText": "query importedCharacterQuery {\n  Character(id: 80) {\n    name {\n      full\n      native\n    }\n  }\n}\n",
